--- src/main/java17/org/codehaus/plexus/archivers/config/FileSetSpec.java
@@ -1,139 +1,0 @@
-/*
- * Copyright MojoHaus and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.codehaus.plexus.archivers.config;
-
-import java.nio.file.Path;
-import java.util.Collection;
-import java.util.List;
-import java.util.Objects;
-
-import org.codehaus.plexus.archiver.FileSet;
-import org.codehaus.plexus.archiver.util.DefaultFileSet;
-import org.codehaus.plexus.components.io.filemappers.FileMapper;
-import org.codehaus.plexus.components.io.fileselectors.FileSelector;
-import org.codehaus.plexus.components.io.functions.InputStreamTransformer;
-
-/**
- * Fluent specification of files rooted at a directory.
- *
- * @since 5.0.0
- */
-public final class FileSetSpec {
-    private final Path directory;
-    private String prefix;
-    private Collection<String> includes;
-    private Collection<String> excludes;
-    private CaseSensitivity caseSensitivity = CaseSensitivity.SENSITIVE;
-    private DefaultExcludes defaultExcludes = DefaultExcludes.USE;
-    private EmptyDirectoryHandling emptyDirectoryHandling = EmptyDirectoryHandling.INCLUDE;
-    private SymbolicLinkHandling symbolicLinkHandling = SymbolicLinkHandling.PRESERVE;
-    private Collection<FileSelector> fileSelectors;
-    private InputStreamTransformer streamTransformer;
-    private Collection<FileMapper> fileMappers;
-
-    private FileSetSpec(Path directory) {
-        this.directory = Objects.requireNonNull(directory, "directory");
-    }
-
-    public static FileSetSpec of(Path directory) {
-        return new FileSetSpec(directory);
-    }
-
-    public FileSetSpec prefixed(String prefix) {
-        this.prefix = prefix;
-        return this;
-    }
-
-    public FileSetSpec including(Collection<String> includes) {
-        this.includes = List.copyOf(includes);
-        return this;
-    }
-
-    public FileSetSpec excluding(Collection<String> excludes) {
-        this.excludes = List.copyOf(excludes);
-        return this;
-    }
-
-    public FileSetSpec caseSensitive(CaseSensitivity caseSensitivity) {
-        this.caseSensitivity = Objects.requireNonNull(caseSensitivity, "caseSensitivity");
-        return this;
-    }
-
-    public FileSetSpec usingDefaultExcludes(DefaultExcludes defaultExcludes) {
-        this.defaultExcludes = Objects.requireNonNull(defaultExcludes, "defaultExcludes");
-        return this;
-    }
-
-    public FileSetSpec emptyDirectories(EmptyDirectoryHandling emptyDirectoryHandling) {
-        this.emptyDirectoryHandling = Objects.requireNonNull(emptyDirectoryHandling, "emptyDirectoryHandling");
-        return this;
-    }
-
-    public FileSetSpec symbolicLinks(SymbolicLinkHandling symbolicLinkHandling) {
-        this.symbolicLinkHandling = Objects.requireNonNull(symbolicLinkHandling, "symbolicLinkHandling");
-        return this;
-    }
-
-    public FileSetSpec selectedBy(Collection<FileSelector> fileSelectors) {
-        this.fileSelectors = List.copyOf(fileSelectors);
-        return this;
-    }
-
-    public FileSetSpec transformedBy(InputStreamTransformer streamTransformer) {
-        this.streamTransformer = Objects.requireNonNull(streamTransformer, "streamTransformer");
-        return this;
-    }
-
-    public FileSetSpec mappedBy(Collection<FileMapper> fileMappers) {
-        this.fileMappers = List.copyOf(fileMappers);
-        return this;
-    }
-
-    FileSet toFileSet() {
-        DefaultFileSet fileSet = new DefaultFileSet(directory.toFile());
-        fileSet.setPrefix(prefix);
-        fileSet.setIncludes(includes == null ? null : includes.toArray(String[]::new));
-        fileSet.setExcludes(excludes == null ? null : excludes.toArray(String[]::new));
-        fileSet.setCaseSensitive(isCaseSensitive());
-        fileSet.setUsingDefaultExcludes(usesBuiltInDefaultExcludes());
-        fileSet.setIncludingEmptyDirectories(includesEmptyDirectories());
-        fileSet.setFollowingSymLinks(followsSymbolicLinks());
-        fileSet.setFileSelectors(fileSelectors == null ? null : fileSelectors.toArray(FileSelector[]::new));
-        if (streamTransformer != null) {
-            fileSet.setStreamTransformer(streamTransformer);
-        }
-        fileSet.setFileMappers(fileMappers == null ? null : fileMappers.toArray(FileMapper[]::new));
-        return fileSet;
-    }
-
-    private boolean isCaseSensitive() {
-        return CaseSensitivities.resolve(caseSensitivity);
-    }
-
-    private boolean usesBuiltInDefaultExcludes() {
-        return defaultExcludes instanceof BuiltInDefaultExcludes builtIn && builtIn.use;
-    }
-
-    private boolean followsSymbolicLinks() {
-        return ((FixedSymbolicLinkHandling) symbolicLinkHandling).following;
-    }
-
-    private boolean includesEmptyDirectories() {
-        return ((FixedEmptyDirectoryHandling) emptyDirectoryHandling).included;
-    }
-}
